+++ packages/deployment/upgrade-test/upgrade-test-scripts/lib/commonUpgradeHelpers.js
@@ -2,7 +2,7 @@
 import { $ } from 'execa';
 import { promises as fs } from 'fs';
 import * as path from 'path';
-import { agd, agoric, agops } from './lib/cliHelper.js';
+import { agd, agoric, agops } from './cliHelper.js';
 import { CHAINID, VALIDATORADDR } from './constants.js';
 
 const waitForBootstrap = async () => {
